feat: 계층형 캐시 도입 및 Redis 안정성 강화 (#86)
- Caffeine(L1)과 Redis(L2)를 조합한 다중 계층 캐시 구조 구현
- Jackson 역직렬화 시 Integer 캐스팅 에러 방지를 위해 반환 타입을 Double로 변경
- Docker-compose 내 Redis maxmemory 제한 및 LRU 방출 정책 설정 추가
- AOP 프록시 환경에서의 동시성 테스트 코드 정상화

+++ src/test/java/maple/expectation/service/CubeServiceTest.java
@@ -35,7 +35,7 @@ class CubeServiceTest {
                 .build();
 
         // 2. when: 이제 '비용'이 아닌 '횟수(Trials)'를 직접 계산합니다.
-        long trials = cubeTrialsProvider.calculateExpectedTrials(input, CubeType.BLACK);
+        long trials = cubeTrialsProvider.calculateExpectedTrials(input, CubeType.BLACK).longValue();
 
         // 3. then
         assertThat(trials).isGreaterThan(0);
@@ -61,7 +61,7 @@ class CubeServiceTest {
                 .build();
 
         // when
-        long trials = cubeTrialsProvider.calculateExpectedTrials(input, CubeType.BLACK);
+        long trials = cubeTrialsProvider.calculateExpectedTrials(input, CubeType.BLACK).longValue();
 
         // then
         assertThat(trials).isGreaterThan(0);
